fix!: carry monetary values as Decimal, not float64
openapi.yml documents Price.value and Price.deposit.value as `type:
number`, but the API actually encodes them as JSON strings (e.g.
"1240.00") to avoid floating-point rounding on currency values — a
spec/wire mismatch confirmed against real API behavior, not something
inferred from the spec itself.

- New Decimal type (decimal.go): a defined string type carrying the exact
  decimal text the API sends, with no arithmetic of its own (this client
  has no runtime dependencies to build one on). UnmarshalJSON also accepts
  a bare JSON number defensively, keeping its literal token rather than
  round-tripping through float64, in case a future spec fix or an
  inconsistency between endpoints ever sends one unquoted.
- Price.Value, PriceDeposit.Value, UpdatePriceParams.Value and
  UpdatePriceParams.Deposit change from float64 to Decimal/
  Nullable[Decimal].

BREAKING CHANGE: Price.Value and PriceDeposit.Value are now
tourmanager.Decimal instead of float64; UpdatePriceParams.Value and
.Deposit change accordingly. No git tag has been cut yet, so this ships
free of an apidiff concern, but is flagged here per the update skill's
breaking-change process for the record.

Audited every `type: number` field across openapi.yml for the same issue:
the only other monetary-shaped fields (searchHolidays min/max price
params, Organization.deposit_defaults) belong to unimplemented resources,
so no other code is affected today. NumberFieldValue (custom field data)
is deliberately left as float64 — it's a generic numeric field, not
specifically monetary.

README.md and AGENTS.md document the Decimal convention.

+++ prices.go
@@ -13,7 +13,7 @@ type PriceDeposit struct {
 	// Calculated reports whether Value was derived from a global deposit
 	// rule, as opposed to being set explicitly on this Price.
 	Calculated bool    `json:"calculated"`
-	Value      float64 `json:"value"`
+	Value      Decimal `json:"value"`
 }
 
 // Price is the cost of booking a single [ElementOption] on a specific
@@ -38,8 +38,10 @@ type Price struct {
 	// field — it's still listed under `properties`). bool is inferred from
 	// its description text ("Determines if the price has been initialized
 	// with a value").
-	Initialized bool    `json:"initialized"`
-	Value       float64 `json:"value"`
+	Initialized bool `json:"initialized"`
+	// Value is this Price's monetary amount, carried as a [Decimal] rather
+	// than a float64 — see Decimal's doc comment for why.
+	Value Decimal `json:"value"`
 
 	// Deposit is nil if no deposit rule applies to this Price.
 	Deposit *PriceDeposit `json:"deposit,omitempty"`
@@ -48,14 +50,17 @@ type Price struct {
 // UpdatePriceParams are the properties that may change on an existing
 // Price. All fields are optional; a nil field is left unchanged.
 //
-// Deposit is `nullable: true` in openapi.yml: use [NullValue] to set an
-// explicit deposit, [Null] to clear it (reverting to the operator's default
-// calculation), or leave it nil to leave the deposit unchanged.
+// Value and Deposit are [Decimal] rather than float64, matching the API's
+// actual (if undocumented) string encoding of monetary values — see
+// Decimal's doc comment. Deposit is `nullable: true` in openapi.yml: use
+// [NullValue] to set an explicit deposit, [Null] to clear it (reverting to
+// the operator's default calculation), or leave it nil to leave the
+// deposit unchanged.
 //
 // spec: updatePrice
 type UpdatePriceParams struct {
-	Value   *float64           `json:"value,omitempty"`
-	Deposit *Nullable[float64] `json:"deposit,omitempty"`
+	Value   *Decimal           `json:"value,omitempty"`
+	Deposit *Nullable[Decimal] `json:"deposit,omitempty"`
 	OnSale  *bool              `json:"on_sale,omitempty"`
 }
